+++ sources/Services.Common/Hub/Quality/IHubQualityCallback.cs
@@ -5,6 +5,6 @@ namespace Queue.Services.Common
     public interface IHubQualityCallback
     {
         [OperationContract(IsOneWay = true)]
-        void Accepted(int rating);
+        void Accepted(byte rating);
     }
 }
